+++ packages/react/src/components/TideGraph/index.ts
@@ -1,6 +1,4 @@
 export * from "./TideGraph.js";
 export * from "./TideGraphChart.js";
-export * from "./TideGraphScroll.js";
-export * from "./TideGraphStatic.js";
 export * from "./NightBands.js";
 export * from "./YAxisOverlay.js";
